+++ packages/strum-react/src/components/Checkbox/Checkbox.tsx
@@ -3,29 +3,30 @@ import { CheckIcon } from '@radix-ui/react-icons';
 import * as RadixLabel from '@radix-ui/react-label';
 import * as React from 'react';
 import { Box } from '../Box';
+import { Text } from '../Text';
 import * as styles from './Checkbox.css';
 
 type CheckboxProps = {
+  /** The checked state of a controlled input */
+  checked?: RadixCheckbox.CheckboxProps['checked'];
   /** Ties the label to the input field */
   id: string;
   /** Description of the field */
   label: string;
   /** onChange callback for a controlled input */
-  onChange?: RadixCheckbox.CheckboxProps['onCheckedChange'];
-  /** The value of a controlled input */
-  value?: RadixCheckbox.CheckboxProps['checked'];
+  onCheckedChange?: RadixCheckbox.CheckboxProps['onCheckedChange'];
 } & Omit<RadixCheckbox.CheckboxProps, 'id' | 'onChange' | 'value'>;
 
 export const Checkbox = React.forwardRef<HTMLButtonElement, CheckboxProps>(
   (
     {
+      checked,
       defaultChecked,
       defaultValue,
       disabled = false,
       id,
       label,
-      onChange,
-      value,
+      onCheckedChange,
       ...primitiveProps
     },
     ref,
@@ -33,13 +34,13 @@ export const Checkbox = React.forwardRef<HTMLButtonElement, CheckboxProps>(
     return (
       <Box alignItems="center" display="flex" gap="4">
         <RadixCheckbox.Checkbox
-          checked={value}
+          checked={checked}
           className={styles.checkboxStyle}
           defaultChecked={defaultChecked}
           defaultValue={defaultValue}
           disabled={disabled}
           id={id}
-          onCheckedChange={onChange}
+          onCheckedChange={onCheckedChange}
           ref={ref}
           {...primitiveProps}
         >
@@ -50,7 +51,7 @@ export const Checkbox = React.forwardRef<HTMLButtonElement, CheckboxProps>(
           </RadixCheckbox.CheckboxIndicator>
         </RadixCheckbox.Checkbox>
         <RadixLabel.Root className={styles.checkboxLabelStyle} htmlFor={id}>
-          {label}
+          <Text color={disabled ? 'neutral11' : 'neutral12'}>{label}</Text>
         </RadixLabel.Root>
       </Box>
     );